added gui and modified account to return int instead of double.

+++ src/GameComponents/Account.java
@@ -4,7 +4,7 @@ import java.io.BufferedReader;
 import java.io.FileReader;
 
 public class Account {
-    private double balance;
+    private int balance;
     private String Accountname;
 
 
@@ -12,20 +12,20 @@ public class Account {
         balance = 0;
         Accountname = name;
     }
-    public void setBalance(double initialBalance){
+    public void setBalance(int initialBalance){
         balance = initialBalance;
     }
-    public void deposit(double amount)
+    public void deposit(int amount)
     {
         balance += amount;
     }
-    public void withDraw(double amount)
+    public void withDraw(int amount)
     {
         balance -= amount;
         if (balance < 0)
             balance = 0;
     }
-    public double getBalance()
+    public int getBalance()
     {
         //System.out.println(Accountname + "now has " + balance + " points");
         return(balance);
